backend/requirements.txt
#requirement imports for containerisation and system running
Flask==2.2.5
Flask_Cors==4.0.0
flask_sqlalchemy==3.1.1
python-dotenv
